use wait_for_connection instead of wait_for with ssh
The check using `wait_for` to check the ssh port does not work
reliably with qemu based testing, because the IP address inside
the VM is not always resolvable externally, and the port number
is different from `22` as well.  Instead, use the `wait_for_connection`
module which explicitly tests that Ansible can connection to the
host using its transport mechanism, which I think is what this
check is trying to do.