use custom vpn_ipaddr filter
Use a custom vpn_ipaddr filter which uses `ipaddress` instead of
`netaddr`.  `ipaddress` is built-in to python3.  On python 2 systems,
the user must install `ipaddress` on the controller.  This is
available as `python-ipaddress` on all EL7 systems.
This is based on the ansible.netcommon.ipaddr filter, but with most
of the functionality removed except that which is necessary for
the VPN role.