System Roles should consistently use ansible_managed in configuration files it manages
bz#2044640

Tests need to support multiline by adding "multiline=True".

Create policy files private, clear, private-or-clear in
/etc/ipsec.d/policies/ from a template policy.j2 for the
entire file, which has the {{ ansible_managed | comment }}
header in it.

In tests_host_to_*.yml, to check the contents of secret files,
instead of comparing the entire content with the expected line,
examine if the expected line exists in the secret file to ignore
the added ansible managed comment.

Signed-off-by: Noriko Hosoi <nhosoi@redhat.com>