docs: add Collection requirements section to README
For roles that use external collections, they have a
`meta/collection-requirements.yml` file, and the README lists
these requirements and instructions.  However, these are only
required for github and Galaxy users, and not for users who
use the packaged roles, where it is confusing to see instructions
about installing requirements when none are needed.  We need to
make it easier to remove these sections when packaging.

Fix some markdownlint issues

Signed-off-by: Rich Megginson <rmeggins@redhat.com>